+++ asyncutils/constants.py
@@ -51,6 +51,5 @@ class _sentinel(sentinel_base):
     def __init_subclass__(cls, /, **_): raise TypeError('cannot subclass the type of asyncutils-internal sentinels')
     def __reduce__(self): return self.name
 _NO_DEFAULT, RAISE = map(_sentinel, ('_NO_DEFAULT', 'RAISE'))
-CLOSED, HALF_OPEN, OPEN = range(3)
 _sentinel._can_instantiate = False
 del _sentinel, P